Updated for 64-bit compatibility

+++ CCMenuTests/Classes/CCMConnectionTest.m
@@ -20,7 +20,7 @@
     connectionMock = [OCMockObject partialMockForObject:connection];
 }
 
-- (id)responseMockWithStatusCode:(int)statusCode
+- (id)responseMockWithStatusCode:(NSInteger)statusCode
 {
     id responseMock = [OCMockObject mockForClass:[NSHTTPURLResponse class]];
     [[[responseMock stub] andReturnValue:OCMOCK_VALUE(statusCode)] statusCode];
